requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# LangChain & AI
langchain>=0.1.0
langchain-groq>=0.0.1
langgraph>=0.0.20

# Twilio for emergency calls
twilio>=8.10.0
